+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## `master`
 
+- Break: use TOML instead of YAML for configuration
 - New: initial Telegram support #18
 - New: `Condition::Or` is able to accept any number of child conditions
 - Chore: `cargo update`